refactor: 컴포넌트 병합, 폴더 구조 변환
