--- components/AllArticleSection.module.css
@@ -1,85 +1,0 @@
-.articleContainer {
-  margin: 24px auto 16px;
-}
-
-.articleHeader {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  width: 1200px;
-  margin-bottom: 24px;
-}
-
-.contentWrapper {
-  display: flex;
-  justify-content: space-between;
-  margin-bottom: 16px;
-}
-
-.allArticleList {
-  display: flex;
-  flex-direction: column;
-  gap: 24px;
-}
-
-.articleTitle {
-  font-size: 20px;
-  font-weight: 700;
-  line-height: 24px;
-  color: #111827;
-}
-
-.articleImageWrapper {
-  position: relative;
-  width: 72px;
-  height: 72px;
-  padding: 10px;
-  border: 1px solid #e5e7eb;
-  border-radius: 6px;
-}
-
-.articleContentBottom {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-bottom: 24px;
-}
-
-.likeCountWrapper {
-  display: flex;
-  align-items: center;
-  gap: 4px;
-}
-
-.writeButton {
-  width: 88px;
-  height: 42px;
-  background-color: #3692ff;
-  border: none;
-  border-radius: 8px;
-  font-size: 16px;
-  font-weight: 600;
-  line-height: 19px;
-  color: #fff;
-}
-
-.allArticleWrapper {
-  display: flex;
-  flex-direction: column;
-  width: 1200px;
-  height: 136px;
-  background-color: #fff;
-  border-bottom: 1px solid #e5e7eb;
-  border-radius: 1px;
-}
-
-.allArticleBottomLeft {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-}
-
-.allArticleSearchBar {
-  width: 1054px;
-  height: 42px;
-}
