--- components/BestArticleSection.module.css
@@ -1,75 +1,0 @@
-.articleContainer {
-  margin: 24px auto 16px;
-}
-
-.articleHeader {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  width: 1200px;
-  margin-bottom: 24px;
-}
-
-.contentWrapper{
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.articleTitle {
-  font-size: 20px;
-  font-weight: 700;
-  line-height: 24px;
-  color: #111827;
-}
-.articleImageWrapper {
-  position: relative;
-  width: 72px;
-  height: 72px;
-  padding: 10px;
-  border: 1px solid #e5e7eb;
-  border-radius: 6px;
-}
-
-.articleContentBottom {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.likeCountWrapper {
-  display: flex;
-  align-items: center;
-  gap: 4px;
-}
-
-.bestArticleList {
-  display: flex;
-  gap: 24px;
-  margin-bottom: 40px;
-}
-
-.bestArticleWrapper {
-  position: relative;
-  width: 384px;
-  height: 169px;
-  background-color: #f9fafb;
-  border-radius: 8px;
-}
-
-.bestArticleContent {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  position: absolute;
-  top: 46px;
-  left: 24px;
-  width: 330px;
-  height: 100px;
-}
-
-.bestArticleBottomLeft {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-}
