--- components/BestProductSection.module.css
@@ -1,22 +1,0 @@
-.bestProductsContainer {
-  width: 1200px;
-  margin: 24px 0 40px;
-}
-
-.bestProductTitle {
-  margin-bottom: 16px;
-  font-size: 20px;
-  font-weight: 700;
-  line-height: 28px;
-  color: #111827;
-}
-
-.bestProductsList {
-  display: flex;
-  gap: 24px;
-}
-
-.bestProductImage {
-  width: 282px;
-  height: 282px;
-}
