+++ components/layout/Container/Container.module.css
@@ -2,6 +2,5 @@
   display: flex;
   flex-direction: column;
   align-items: center;
-  /* margin-top: 70px; */
 }
