+++ components/ui/DropDownButton/DropDownButton.module.css
@@ -1,8 +1,8 @@
-.dropDownButtonWrapper{
+.dropDownButtonWrapper {
   position: relative;
 }
 
-.dropDownButton{
+.dropDownButton {
   display: flex;
   justify-content: center;
   align-items: center;
@@ -11,11 +11,11 @@
   height: 42px;
   background-color: #fff;
   cursor: pointer;
-  border: 1px solid #e5e7eb;
+  border: 1px solid var(--color-gray200);
   border-radius: 12px;
 }
 
-.dropDownMenu{
+.dropDownMenu {
   display: flex;
   flex-direction: column;
   position: absolute;
@@ -23,25 +23,25 @@
   top: 50px;
   left: 0px;
   background-color: #fff;
-  border: 1px solid#e5e7eb;
+  border: 1px solid var(--color-gray200);
   border-radius: 8px;
 }
 
-.dropDownItem{
+.dropDownItem {
   display: flex;
   justify-content: center;
   align-items: center;
   width: 130px;
   height: 42px;
-  border-bottom: 1px solid #e5e7eb;
+  border-bottom: 1px solid var(--color-gray200);
   font-size: 16px;
 }
 
-.dropDownItem:last-child{
+.dropDownItem:last-child {
   border-bottom: none;
 }
 
-.dropDownItem:hover{
-  background-color: #e5e7eb;
+.dropDownItem:hover {
+  background-color: var(--color-gray200);
   cursor: pointer;
 }
