+++ components/ui/SearchForm/SearchForm.module.css
@@ -1,17 +1,17 @@
-.searchInputWrapper{
+.searchInputWrapper {
   position: relative;
 }
 
 .searchInput {
   padding: 9px 44px;
   background-color: #f3f4f6;
-  border: none;
+
   border-radius: 12px;
   font-size: 16px;
   font-weight: 400;
   line-height: 24px;
 }
 
-.searchInput:focus{
-  outline: 1px solid #3692ff;
+.searchInput:focus {
+  outline: 1px solid var(--color-blue);
 }
