+++ components/ui/SearchForm/SearchForm.tsx
@@ -1,6 +1,6 @@
 import Image from "next/image";
 import Search from "@/public/images/icons/ic_search.svg";
-import styles from "@/components/SearchForm.module.css";
+import styles from "@/components/ui/SearchForm/SearchForm.module.css";
 
 interface SearchFormProps {
   onInputChange: (event: string) => void;
