+++ pages/HomePage.module.css
@@ -23,17 +23,17 @@
 .top_banner_button {
   width: 357px;
   height: 56px;
-  background-color: #3692ff;
-  border: none;
+  background-color: var(--color-blue);
+
   border-radius: 40px;
   font-size: 20px;
   font-weight: 600;
   line-height: 24px;
   color: #fff;
   text-align: center;
-} 
+}
 
-.top_banner_button:hover{
+.top_banner_button:hover {
   background-color: #1251aa;
   cursor: pointer;
 }
@@ -77,7 +77,7 @@
   font-size: 18px;
   font-weight: 700;
   line-height: 25px;
-  color: #3692ff;
+  color: var(--color-blue);
 }
 
 .section h1 {
@@ -127,7 +127,8 @@
   background-color: #111322;
 }
 
-.page_info, .important_info{
+.page_info,
+.important_info {
   font-size: 16px;
   font-weight: 400;
   line-height: 20px;
@@ -148,7 +149,7 @@
   gap: 12px;
 }
 
-.fotter_icons img{
+.fotter_icons img {
   width: 20px;
   height: 20px;
 }
