+++ pages/additem/AddItemPage.module.css
@@ -16,14 +16,14 @@
   font-size: 28px;
   font-weight: 700;
   line-height: 33px;
-  color: #1f2937;
+  color: var(--color-gray800);
 }
 
 .addButton {
   width: 88px;
   height: 42px;
-  background-color: #3692ff;
-  border: none;
+  background-color: var(--color-blue);
+
   border-radius: 8px;
   font-size: 16px;
   font-weight: 600;
@@ -32,7 +32,7 @@
 }
 
 .addButton:disabled {
-  background-color: #9ca3af;
+  background-color: var(--color-gray400);
 }
 
 .addFormWrapperLabel {
@@ -47,12 +47,12 @@
   margin: 12px 0 24px 0;
   padding: 16px 24px;
   background-color: #f3f4f6;
-  border: none;
+
   border-radius: 12px;
   font-size: 16px;
   font-weight: 400;
   line-height: 24px;
-  color: #9ca3af;
+  color: var(--color-gray400);
 }
 
 .addFormWrapper textarea {
@@ -78,15 +78,15 @@
   border-radius: 26px;
   font-size: 16px;
   font-weight: 400;
-  color: #1f2937;
+  color: var(--color-gray800);
 }
 
 .tagWrapper ul li button {
   position: relative;
   width: 20px;
   height: 20px;
-  background-color: #9ca3af;
-  border: none;
+  background-color: var(--color-gray400);
+
   border-radius: 999px;
   color: #fff;
 }
@@ -99,6 +99,6 @@
 }
 
 .tagWrapper ul li button:hover {
-  background-color: #3692ff;
+  background-color: var(--color-blue);
   cursor: pointer;
 }
