+++ pages/additem/index.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useRef, useState } from "react";
 import Image from "next/image";
 import Close from "@/public/images/icons/ic_x.svg";
-import styles from "@/styles/AddItemPage.module.css";
-import FileInput from "@/components/FileInput";
+import styles from "@/pages/additem/AddItemPage.module.css";
+import FileInput from "@/components/ui/FileInput/FileInput";
 
 interface InputValue {
   title: string;
