+++ pages/boards/CommunityFeedPage.module.css
@@ -1,0 +1,128 @@
+.articleContainer {
+  margin: 24px auto 16px;
+}
+
+.articleHeader {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  width: 1200px;
+  margin-bottom: 24px;
+}
+
+.articleTitle {
+  font-size: 20px;
+  font-weight: 700;
+  line-height: 24px;
+  color: var(--color-gray900);
+}
+
+.articleImageWrapper {
+  position: relative;
+  width: 72px;
+  height: 72px;
+  padding: 10px;
+  border: 1px solid var(--color-gray200);
+  border-radius: 6px;
+}
+
+.bestContentWrapper {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.bestArticleContentBottom {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.likeCountWrapper {
+  display: flex;
+  align-items: center;
+  gap: 4px;
+}
+
+.bestArticleList {
+  display: flex;
+  gap: 24px;
+  margin-bottom: 40px;
+}
+
+.bestArticleWrapper {
+  position: relative;
+  width: 384px;
+  height: 169px;
+  background-color: #f9fafb;
+  border-radius: 8px;
+}
+
+.bestArticleContent {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  position: absolute;
+  top: 46px;
+  left: 24px;
+  width: 330px;
+  height: 100px;
+}
+
+.bestArticleBottomLeft {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+}
+
+.allContentWrapper {
+  display: flex;
+  justify-content: space-between;
+  margin-bottom: 16px;
+}
+
+.allArticleList {
+  display: flex;
+  flex-direction: column;
+  gap: 24px;
+}
+
+.allArticleContentBottom {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 24px;
+}
+
+.writeButton {
+  width: 88px;
+  height: 42px;
+  background-color: var(--color-blue);
+
+  border-radius: 8px;
+  font-size: 16px;
+  font-weight: 600;
+  line-height: 19px;
+  color: #fff;
+}
+
+.allArticleWrapper {
+  display: flex;
+  flex-direction: column;
+  width: 1200px;
+  height: 136px;
+  background-color: #fff;
+  border-bottom: 1px solid var(--color-gray200);
+  border-radius: 1px;
+}
+
+.allArticleBottomLeft {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+}
+
+.allArticleSearchBar {
+  width: 1054px;
+  height: 42px;
+}
