+++ pages/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import Link from "next/link";
 import Image from "next/image";
-import styles from "@/styles/HomePage.module.css";
+import styles from "@/pages/HomePage.module.css";
 import TopBanner from "@/public/images/home/hero-image.png";
 import Home1 from "@/public/images/home/feature1-image.png";
 import Home2 from "@/public/images/home/feature2-image.png";
