+++ pages/items/MarketPage.module.css
@@ -1,3 +1,26 @@
+.bestProductsContainer {
+  width: 1200px;
+  margin: 24px 0 40px;
+}
+
+.bestProductTitle {
+  margin-bottom: 16px;
+  font-size: 20px;
+  font-weight: 700;
+  line-height: 28px;
+  color: var(--color-gray900);
+}
+
+.bestProductsList {
+  display: flex;
+  gap: 24px;
+}
+
+.bestProductImage {
+  width: 282px;
+  height: 282px;
+}
+
 .allProductContainer {
   width: 1200px;
   margin-bottom: 40px;
@@ -19,14 +42,14 @@
   font-size: 20px;
   font-weight: 700;
   line-height: 28px;
-  color: #111827;
+  color: var(--color-gray900);
 }
 
 .addProductButton {
   width: 133px;
   height: 42px;
-  background-color: #3692ff;
-  border: none;
+  background-color: var(--color-blue);
+
   border-radius: 8px;
   font-size: 16px;
   font-weight: 600;
