+++ styles/SignPage.module.css
@@ -28,7 +28,7 @@
   margin-bottom: 24px;
   padding: 16px 24px;
   background-color: #f3f4f6;
-  border: none;
+
   border-radius: 12px;
   font-size: 16px;
   font-weight: 400;
@@ -43,7 +43,7 @@
   width: 640px;
   height: 56px;
   background-color: #3182f6;
-  border: none;
+
   border-radius: 40px;
   font-size: 20px;
   font-weight: 600;
@@ -51,13 +51,13 @@
   color: #fff;
 }
 
-.sign_form button:hover{
+.sign_form button:hover {
   background-color: #1251aa;
   cursor: pointer;
 }
 
 .sign_form button:disabled {
-  background-color: #9ca3af;
+  background-color: var(--color-gray400);
 }
 
 .easy_login {
@@ -76,7 +76,7 @@
   font-size: 16px;
   font-weight: 500;
   line-height: 24px;
-  color: #1f2937;
+  color: var(--color-gray800);
 }
 
 .login_icons {
@@ -104,5 +104,5 @@
   font-size: 15px;
   font-weight: 600;
   line-height: 18px;
-  color: #f74747;
+  color: var(--color-red);
 }
